fix: FileWatcher adds sub entries if a folder is created
In Windows, DirectoryWatcher fails to emit ENTRY_CREATE events for sub directories and files in those sub directories if a folder is created.
It is known issue in https://github.com/gmethvin/directory-watcher/issues/10
This commit updates FileWatcher so that it walks sub entries of the new folder and adds them to watching set.
This commit fixes #52.